Ported the simplify over from Cura Engine.
With a few adjustments to use a more generic point_container.

This should help with benchmarking the impact of the plugin.
We can test how fast the normal simplify is and compare it
against the plugin (running locally and remote) or the
default behaviour of the plugin.

Contributes to CURA-10475