Add handshake service
The modifications include the addition of a handshake service and some reorganization of
the code. The handshake service was added to enable a preliminary communication process
between the client and the server. This would allow the server to identify the
characteristics of the client before proceeding further.

Contributes to CURA-10714